Refactor and rename title/qid wrappers
- Move Qid and Title to separate modules
- Reformat benchmark
